+++ service/src/main/java/com/epam/esm/service/util/TagValidator.java
@@ -5,6 +5,6 @@ import com.epam.esm.service.exception.ServiceException;
 
 public interface TagValidator {
     void validateTag(Tag tag) throws ServiceException;
-    void validateId(long id) throws ServiceException;
+    void validateId(int id) throws ServiceException;
     void validateName(String name) throws ServiceException;
 }
